To retain compatibviltiy with older cuda architectures I had to sliglhy rearange spline struct. Add SplineCommon whihc are comon defines between cpu and gpu, while SplineStructs contiant actual structs and can use some convoluted c++ templates
